some errors fiexed

+++ src/main/java/com/example/demo/repository/PrescriptionDrugRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.PrescriptionDrug;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PrescriptionDrugRepository extends JpaRepository<PrescriptionDrug, Long> {
+public interface PrescriptionDrugRepository extends JpaRepository<PrescriptionDrug, Integer> {
 }
