+++ src/main/java/com/example/demo/repository/PrescriptionRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.Prescription;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PrescriptionRepository extends JpaRepository<Prescription, Long> {
+public interface PrescriptionRepository extends JpaRepository<Prescription, Integer> {
 }
